Modified for-loops that loop on stl containers to use 'size_t' instead of 'int' for the loop variable, in order to remove 'signed/unsigned integer comparison' compiler warnings. Modified xml code generation tools along the same lines and regenerated all state Attributes and Python attributes which did not contain un-regenerateable specialized code.
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@4154 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/DBAtts/MetaData/avtArrayMetaData.C
@@ -484,7 +484,7 @@ avtArrayMetaData::Print(ostream &out, int indent) const
     out << "Number of variables = " << nVars << endl;
     Indent(out, indent);
     out << "Components are: ";
-    for (int i = 0 ; i < compNames.size() ; i++)
+    for (size_t i = 0 ; i < compNames.size() ; i++)
     {
         out << compNames[i].c_str();
         if (i != compNames.size()-1)
